requirements.txt
PyQt5
pyyaml
doit
https://github.com/pyinstaller/pyinstaller/tarball/develop
